balance method
